Created 12 hour display with 24 hour internals
Created ampm string to hold AM or PM for display of a 12 hour clock.
Updated both constructors to use the ampm String.
Changed timeTick to set the ampm variable when hour changes.
Changed updateDisplay to create a 12 hour display.
Created setAmpm to se the variable based on a 24 hour clock